Remove host-creation endpoint, function, and tests
